db_migrations/20211201_01_Adhx7-create-table-dyndns_updates.py
"""
CREATE TABLE dyndns_updates
"""

from yoyo import step

__depends__ = {}

steps = [
    step("CREATE TABLE dyndns_updates (id INT, type VARCHAR(4), value VARCHAR(128), date VARCHAR(128), PRIMARY KEY (id))")
]
